Move InterruptService to its own module and make the NRF52 chip statically generic over it instead of dynamically.
