+++ chips/nrf52/src/lib.rs
@@ -12,6 +12,7 @@ mod deferred_call_tasks;
 pub mod ficr;
 pub mod i2c;
 pub mod ieee802154_radio;
+pub mod interrupt_service;
 pub mod nvmc;
 pub mod power;
 pub mod ppi;
