+++ chips/nrf52832/src/chip.rs
@@ -1,3 +1,1 @@
 pub use nrf52::chip::NRF52;
-
-pub type Nrf52832InterruptService = nrf52::chip::Nrf52InterruptService;
