+++ chips/nrf52832/src/interrupt_service.rs
@@ -1,0 +1,1 @@
+pub type Nrf52832InterruptService = nrf52::interrupt_service::Nrf52InterruptService;
