+++ chips/nrf52840/src/interrupt_service.rs
@@ -1,14 +1,14 @@
 use crate::peripheral_interrupts;
-use nrf52::chip::InterruptService;
+use nrf52::interrupt_service::InterruptService;
 
 pub struct Nrf52840InterruptService {
-    nrf52: nrf52::chip::Nrf52InterruptService,
+    nrf52: nrf52::interrupt_service::Nrf52InterruptService,
 }
 
 impl Nrf52840InterruptService {
     pub unsafe fn new(gpio_port: &'static nrf52::gpio::Port) -> Nrf52840InterruptService {
         Nrf52840InterruptService {
-            nrf52: nrf52::chip::Nrf52InterruptService::new(gpio_port),
+            nrf52: nrf52::interrupt_service::Nrf52InterruptService::new(gpio_port),
         }
     }
 }
